inline Perehod and extract RoomsReader
начал избавляться от постоянного перечитывания файла Komnaty
Для этого нужно где-то прочитанный файл хранить, поэтому класс Rooms обзавёлся состоянием,
а логика чтения из него выехала в отдельный класс RoomsReader.
Чтобы упростить дальнейшее переиспользование Rooms в других местах,
заинлайнил Perehod в Skasat
